Remove circular dependency for Roles
